Update `useDeltas` -> `useDelta` (#953)
Accept a single delta, so it matches the API for `createStore`.
Both of these accept a single delta. We should intead encourage
a root delta that combines them all together, providing a point
where the application state & actions can be decoupled from
an individual delta's. We could eventually provide a helper for
this process but start with suggesting manual combinations.