Move to docs package & add codesandbox snippets
We'll maintain the code used throughout the walkthrough in the
examples folder. We can also use this to ensure the packages
continue to work with the intended use case.

The codesandbox examples had some isuses when function as part of
the brookjs package. This makes it easier to maintain the docs
separately.