Fix project packaging for subdependencies
This _should_ resolve the problem once and for all. What we're doing
is, prior to packing the package, updating the package's pkg.json
to point to the resulting tarballs instead of a version number. This
way, we don't need to update the package dependencies to point
to anything else and everything should correctly point to our local
packed packages.