Add jsonp config to webpack
This is a big-ish refactor of the `project` & `webpack` modules, and
combining them in the `start` & `build` commands. This provides a
lot more reuse between the two commands, which both interact with
Webpack a lot, so we can abstract more of the reducer logic & config
out to that module, and just make the commands bindings between
the input arguments & the view.

All of this work so that we don't add a lot of duplication to add the
`name` property, which we'll read from the `pkg.json`. This logic
is now in the `project` module.